Move flap command into flap subsystem
