+++ src/main/java/frc/robot/commandgroups/ShootCargo.java
@@ -3,7 +3,7 @@ package frc.robot.commandgroups;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import frc.robot.subsystems.conveyor.Conveyor;
 import frc.robot.subsystems.conveyor.commands.Convey;
-import frc.robot.subsystems.conveyor.commands.FlapCommand;
+import frc.robot.subsystems.flap.commands.FlapCommand;
 import frc.robot.subsystems.flap.Flap;
 import frc.robot.subsystems.hood.Hood;
 import frc.robot.subsystems.hood.commands.HoodCommand;
